D328.WPF/Converter/LineModeToRecordingStopButtonEnabledConverter.cs
using D328.Domain.Enum;
using System;
using System.Globalization;
using System.Windows.Data;

namespace D328.WPF.Converter
{
    public class LineModeToRecordingStopButtonEnabledConverter : IValueConverter
    {
        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            var lineMode = value as LineMode?;
            if (lineMode == null)
            {
                return false;
            }
            return lineMode == LineMode.Recording || lineMode == LineMode.Pause;
        }

        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
        {
            throw new NotImplementedException();
        }
    }
}
